requirements.txt
mss>=9.0,<10.0
pynput>=1.7,<2.0
pystray>=0.19,<0.20
Pillow>=10.0,<12.0
